+++ api/src/main/java/com/onion/api/userapi/UserServiceApi.java
@@ -1,6 +1,5 @@
-package com.onion.web.service;
+package com.onion.api.userapi;
 
-import com.onion.api.userapi.UserServiceApi;
 import com.onion.pojo.user.User;
 import org.springframework.cloud.netflix.feign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,7 +10,13 @@ import org.springframework.web.bind.annotation.RequestMethod;
 /**
  * Created by OnionMac on 2018/2/6.
  */
-@FeignClient("user-service")
-public interface UserService extends UserServiceApi {
+@RequestMapping("user")
+public interface UserServiceApi {
+
+    @GetMapping("/getName")
+    String getName();
+
+    @RequestMapping(value = "/getUser/{id}",method = RequestMethod.GET)
+    User findUserById(@PathVariable("id") int id);
 
 }
