+++ userservice/src/main/java/com/onion/userservice/controller/UserController.java
@@ -1,7 +1,8 @@
 package com.onion.userservice.controller;
 
-import com.onion.api.userapi.UserService;
+import com.onion.api.userapi.UserServiceApi;
 import com.onion.pojo.user.User;
+import com.onion.userservice.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,19 +13,19 @@ import org.springframework.web.bind.annotation.RestController;
  * Created by OnionMac on 2018/2/6.
  */
 @RestController
-@RequestMapping("user")
-public class UserController {
+public class UserController implements UserServiceApi {
 
     @Autowired
     UserService userService;
 
-    @GetMapping("/getName")
+    @Override
     public String getName(){
         return userService.getName();
     }
 
-    @GetMapping("/getUser/{id}")
-    public User getUser(@PathVariable("id") int id){
+    @Override
+    public User findUserById(int id) {
         return userService.findUserById(id);
     }
+
 }
