+++ web/src/main/resources/application.yml
@@ -10,7 +10,7 @@ server:
     mime-types: application/json,application/xml,text/html,text/xml,text/plain
 eureka:
   service-url:
-      defaultZone: http://peer1:8761/eureka/,http://peer2:8762/eureka/  #注册到eureka
+      defaultZone: http://localhost:8761/eureka/  #注册到eureka
   instance:
     prefer-ip-address: true
